requirements.txt
telethon
tqdm
loguru
retrying
python-dotenv
